+++ src/com/kaylerrenslow/armaDialogCreator/data/PaaImageExternalResource.java
@@ -2,7 +2,6 @@ package com.kaylerrenslow.armaDialogCreator.data;
 
 import com.kaylerrenslow.armaDialogCreator.util.KeyValueString;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 import java.io.File;
 
@@ -13,33 +12,18 @@ import java.io.File;
  @since 07/19/2016. */
 public class PaaImageExternalResource extends ExternalResource {
 	/**
-	 Key used for the original .paa image. To get the .paa image file path,
+	 Key used for the converted .paa image. To get the converted image file path,
 	 use key with {@link #getPropertyValue(String)}
 	 */
-	public static final String keyPaaImagePath = "paa-image-path";
+	public static final String KEY_CONVERTED_IMAGE = "converted-paa";
 
 	/**
 	 Create a link for a .paa image
 
-	 @param paaImagePath path to the .paa image that was used
+	 @param paaImage file to the .paa image that was used
 	 @param convertedImage image that was converted
 	 */
-	public PaaImageExternalResource(@NotNull File paaImagePath, @NotNull File convertedImage) {
-		super(convertedImage, new KeyValueString[]{new KeyValueString(keyPaaImagePath, paaImagePath.getPath())});
-	}
-
-	@Nullable
-	public static File getPaaImagePath(@NotNull ExternalResource resource) {
-		KeyValueString kv = resource.getPropertyValue(keyPaaImagePath);
-		if (kv == null) {
-			return null;
-		}
-		return new File(kv.getValue());
-	}
-
-	/** Get the path to the converted .paa image (stored as a .png) */
-	@Nullable
-	public File getPaaImagePath() {
-		return getPaaImagePath(this);
+	public PaaImageExternalResource(@NotNull File paaImage, @NotNull File convertedImage) {
+		super(paaImage, new KeyValueString[]{new KeyValueString(KEY_CONVERTED_IMAGE, convertedImage.getAbsolutePath())});
 	}
 }